Project structure + minor code refactor. Fixed the double displaying swipeable view. Other small fixes. README update.
