+++ Client/bisMateClient/bisMateClient/OutboundRequesting/Clients/WSocketClient/SocketClient.swift
